Fix bug: Equipped don't equip on loading save
When loading from a saved game, any item you had equipped would not be equipped when loading the game. This happened because
before saving, the item in your inventory and the item in your hand would be the same instance of Item. After loading, they would be
two different instances of Item.